requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17.2
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
chromedriver-binary==96.0.4664.45.0
click==8.0.3
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
cycler==0.11.0
discord.py==1.7.3
Flask==2.0.2
fonttools==4.28.5
future==0.18.2
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
lxml==4.7.1
MarkupSafe==2.0.1
matlab==0.1
matplotlib==3.5.1
multidict==6.0.2
natsort==8.0.2
numpy==1.21.5
oauthlib==3.2.0
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
outcome==1.1.0
packaging==21.3
pandas==1.3.5
pefile==2021.9.3
Pillow==8.4.0
pycodestyle==2.8.0
pycparser==2.21
pyinstaller==4.8
pyinstaller-hooks-contrib==2022.0
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
pywin32-ctypes==0.2.0
requests==2.27.0
requests-oauthlib==1.3.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
tweepy==4.5.0
typing_extensions==4.0.1
urllib3==1.26.7
webdriver-manager==3.5.2
Werkzeug==2.0.2
wsproto==1.0.0
yarl==1.7.2
youtube-dl==2021.12.17
